Refactor generative research maintenance hotspots
